requirements.txt
# ============================================
# Lymph-Node - Requisitos del Proyecto
# Python 3.11 requerido
# Instala las versiones más actuales disponibles
# ============================================

# === Core: Procesamiento de Imágenes Médicas ===
tcia-utils
SimpleITK

# === Deep Learning: Segmentación 3D ===
tensorflow

# === Radiomics: Extracción de Características ===
scikit-image

# === Machine Learning: Clasificación ===
scikit-learn
xgboost
imbalanced-learn

# === Data Science: Análisis y Manipulación ===
pandas
numpy
scipy

# === Visualización: Gráficos y Dashboard ===
matplotlib
seaborn
plotly

# === Utilidades: Métricas y Evaluación ===
opencv-python

# === Utilidades: Radiomicas ===
pyradiomics   

# === Utilidades: Progreso y Manejo de Archivos ===
tqdm